test: split tests into multiple files
The `test_advanced.py` file had over 1000 lines.
